+++ MyModel2.cpp
@@ -18,7 +18,7 @@ MyModel2::MyModel2()
 ,ys(Data::get_instance().get_y_rays())
 ,surface_brightness(Data::get_instance().get_x_rays())
 ,model_image(Data::get_instance().get_ni(),
-		vector<long double>(Data::get_instance().get_nj()))
+		vector<double>(Data::get_instance().get_nj()))
 {
 
 }
@@ -111,8 +111,8 @@ void MyModel2::print(std::ostream& out) const
 	source.print(out);
 
 	// Make an image of the source (uses the ray resolution)
-	const vector< vector<long double> >& x = Data::get_instance().get_x_rays();
-	const vector< vector<long double> >& y = Data::get_instance().get_y_rays();
+	const vector< vector<double> >& x = Data::get_instance().get_x_rays();
+	const vector< vector<double> >& y = Data::get_instance().get_y_rays();
 	for(size_t i=0; i<xs.size(); i++)
 		for(size_t j=0; j<xs[i].size(); j++)
 			out<<source.evaluate(x[i][j], y[i][j])<<' ';
@@ -135,8 +135,8 @@ string MyModel2::description() const
 
 void MyModel2::shoot_rays()
 {
-	const vector< vector<long double> >& x = Data::get_instance().get_x_rays();
-	const vector< vector<long double> >& y = Data::get_instance().get_y_rays();
+	const vector< vector<double> >& x = Data::get_instance().get_x_rays();
+	const vector< vector<double> >& y = Data::get_instance().get_y_rays();
 	const vector< vector<double> >& sigma = Data::get_instance().get_sigma();
 
 	double ax, ay;
@@ -189,8 +189,8 @@ void MyModel2::update_rays()
 {
 	staleness2++;
 
-	const vector< vector<long double> >& x = Data::get_instance().get_x_rays();
-	const vector< vector<long double> >& y = Data::get_instance().get_y_rays();
+	const vector< vector<double> >& x = Data::get_instance().get_x_rays();
+	const vector< vector<double> >& y = Data::get_instance().get_y_rays();
 
 	double ax, ay;
 	for(size_t i=0; i<xs.size(); i++)
@@ -209,7 +209,7 @@ void MyModel2::calculate_model_image()
 	int resolution = Data::get_instance().get_resolution();
 
 	model_image.assign(Data::get_instance().get_ni(),
-		vector<long double>(Data::get_instance().get_nj(), 0.));
+		vector<double>(Data::get_instance().get_nj(), 0.));
 
 	int ii, jj;
 	double coeff = pow(static_cast<double>(resolution), -2);
